fix p0
