+++ fe/fe-core/src/main/java/org/apache/doris/qe/SessionVariable.java
@@ -1848,7 +1848,7 @@ public class SessionVariable implements Serializable, Writable {
     public int createTablePartitionMaxNum = 10000;
 
 
-    @VariableMgr.VarAttr(name = KEEP_CARRIAGE_RETURN, needForward = true,
+    @VariableMgr.VarAttr(name = KEEP_CARRIAGE_RETURN,
             description = {"在同时处理\r和\r\n作为CSV的行分隔符时，是否保留\r",
                     "When processing both \\n and \\r\\n as CSV line separators, should \\r be retained?"})
     public boolean keepCarriageReturn = false;
