--- regression-test/data/external_table_p0/tvf/test_tvf_csv.out
@@ -1,353 +1,0 @@
--- This file is automatically generated. You should know what you did if you want to edit this
--- !csv_1 --
-1	Alice	30	New York\r
-10	abc	ttt	def
-2	Bob	25	Los Angeles\r
-3	Charlie	35	Chicago\r
-4	abc	def	sss
-5	ttt	def	sss
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-w	w	w	w\r
-wrweqreqer	234	32323	3232
-
--- !csv_2 --
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-2
-
--- !csv_3 --
-1	Alice	30	New York
-10	abc	ttt	def
-2	Bob	25	Los Angeles
-3	Charlie	35	Chicago
-4	abc	def	sss
-5	ttt	def	sss
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-w	w	w	w
-wrweqreqer	234	32323	3232
-
--- !csv_4 --
-
